requirement.txt
streamlit==1.24.3
panda==2.0.3
seaborn==0.13.2
numpy==1.24.3